Wire up querystring for subject
