+++ hs_discover/public/index.html
@@ -42,6 +42,7 @@
     }
 </style>
 <input type="hidden" id="qstring" value="{{ request.GET.q }}">
+<input type="hidden" id="prefillsubject" value="{{ request.GET.subject }}">
 <div class="container" id="discover-main">
     <h2 class="page-title">Discover
         <small class="text-muted"><i>Public resources shared with the community</i></small>
